office_os/pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-office_os"
version = "0.1.0"
description = "Office Os environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.0",
    # Google Sheets integration (real-world dashboard + invoices)
    "gspread>=6.0.0",
    # LLM agent intelligence (Claude API)
    "anthropic>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m office_os.server.app
server = "office_os.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["office_os", "office_os.server", "office_os.market", "office_os.agents", "office_os.integrations"]
package-dir = { "office_os" = ".", "office_os.server" = "server", "office_os.market" = "market", "office_os.agents" = "agents", "office_os.integrations" = "integrations" }